+++ src/AllTables/tickets/ticket.route.ts
@@ -1,13 +1,21 @@
 //routing
 import { Express } from "express";
 import { createTicketController, deleteTicketController, getAllTicketController, getTicketAndUserController, getTicketByIdController, getTicketByUserIdController, updateTicketController } from "./ticket.controller";
+import { allRoleAuth, requireOwnerOrAdmin } from "../../middleware/tokensAuth";
+import { getTicketByIDService } from "./ticket.service";
 
+// Ownership resolver for a support ticket accessed by its own TicketID
+const ticketOwnerResolver = async (req: any) => {
+    const ticket = await getTicketByIDService(parseInt(req.params.id));
+    return ticket?.UserID ?? null;
+}
 
 //CRUD
 const TicketRoutes = (app: Express) => {
     //route
     //Add new Ticket
     app.route("/ticket/newTicket").post(
+        allRoleAuth,
         async (req, res, next) =>{
             try {
                 await createTicketController(req, res);
@@ -30,6 +38,8 @@ const TicketRoutes = (app: Express) => {
 
     //get Ticket by ID
     app.route("/ticket/:id").get(
+        allRoleAuth,
+        requireOwnerOrAdmin(ticketOwnerResolver),
         async (req, res, next) =>{
             try {
                 await getTicketByIdController(req, res);
@@ -41,6 +51,8 @@ const TicketRoutes = (app: Express) => {
 
        //get Ticket by User ID
     app.route("/ticket/user/:id").get(
+        allRoleAuth,
+        requireOwnerOrAdmin(async (req) => parseInt(req.params.id)),
         async (req, res, next) =>{
             try {
                 await getTicketByUserIdController(req, res);
@@ -52,6 +64,8 @@ const TicketRoutes = (app: Express) => {
 
         //get Ticket by User ID
     app.route("/ticket/ticketAndUser/:id").get(
+        allRoleAuth,
+        requireOwnerOrAdmin(async (req) => parseInt(req.params.id)),
         async (req, res, next) =>{
             try {
                 await getTicketAndUserController(req, res);
@@ -63,6 +77,8 @@ const TicketRoutes = (app: Express) => {
 
     //update Ticket by id
     app.route("/ticket/updateticket/:id").patch(
+        allRoleAuth,
+        requireOwnerOrAdmin(ticketOwnerResolver),
         async (req, res, next) => {
             try {
                 await updateTicketController(req, res);
@@ -74,6 +90,8 @@ const TicketRoutes = (app: Express) => {
 
     //Delete Ticket by ID
     app.route("/ticket/delete/:id").delete(
+        allRoleAuth,
+        requireOwnerOrAdmin(ticketOwnerResolver),
         async (req, res, next) =>{
             try {
                 await deleteTicketController(req, res);
